Add Clash.Normalize.Transformations.Specialize
All transformations concerning specializing now live in a separate module.
This includes application propagation as it is used directly by other
specialization transformations. This is part of work to make the
normalization modules quicker to compile.
